rename: utils 파일 생성 및 적용
